// src/app/models/character.ts
export interface Character {
  actor: string;
  alive: boolean;
  ancestry: string;
  dateOfBirth: string;
  eyeColour: string;
  gender: string;
  hairColour: string;
  hogwartsStaff: boolean;
  hogwartsStudent: boolean;
  house: string;
  image: string;
  name: string;
  patronus: string;
  species: string;
  wand: object;
  yearOfBirth: number;
}